+++ examples/11_snake.cpp
@@ -46,13 +46,13 @@ int main()
     bool paused = false;
 
     // Movement timer
-    float moveTimer = 0;
-    float moveInterval = 0.15f;
+    double moveTimer = 0.0;
+    double moveInterval = 0.15;
 
     while (!game.IsClosed()) {
         if (game.IsKeyPressed(KEY_ESCAPE)) break;
 
-        float dt = game.GetDeltaTime();
+        double dt = game.GetDeltaTime();
 
         if (!gameOver) {
             // Pause
@@ -112,7 +112,7 @@ int main()
                             }
                             score += 10;
                             // Speed up
-                            if (moveInterval > 0.06f) moveInterval -= 0.003f;
+                            if (moveInterval > 0.06) moveInterval -= 0.003;
 
                             // New food (ensure not on snake)
                             bool onSnake;
@@ -142,7 +142,7 @@ int main()
                 dir = 3; nextDir = 3;
                 foodR = 5; foodC = 15;
                 score = 0;
-                moveInterval = 0.15f;
+                moveInterval = 0.15;
                 gameOver = false;
             }
         }
