+++ examples/14_tilemap.cpp
@@ -142,14 +142,14 @@ int main()
     int treeY = (FG_R - 3) * TS - 61;
 
     // ---- Player ----
-    float playerX = 48.0f;
-    float playerY = (float)((fgRows - 3) * fgTileSize - CHAR_FOOT); // Stand on ground
-    float speed = 180.0f;
+    double playerX = 48.0;
+    double playerY = (double)((fgRows - 3) * fgTileSize - CHAR_FOOT); // Stand on ground
+    double speed = 180.0;
     int facing = 1;    // 1=right, -1=left
 
     while (!game.IsClosed()) {
-        float dt = game.GetDeltaTime();
-        if (dt > 0.05f) dt = 0.05f; // Prevent first frame jump
+        double dt = game.GetDeltaTime();
+        if (dt > 0.05) dt = 0.05; // Prevent first frame jump
 
         // ---- Input ----
         if (game.IsKeyDown(KEY_RIGHT) || game.IsKeyDown(KEY_D)) {
@@ -162,11 +162,11 @@ int main()
         }
         // Limit player range
         if (playerX < 0) playerX = 0;
-        if (playerX > fgWorldWidth - CHAR_W) playerX = (float)(fgWorldWidth - CHAR_W);
+        if (playerX > fgWorldWidth - CHAR_W) playerX = (double)(fgWorldWidth - CHAR_W);
 
         // ---- Camera follow ----
-        float cameraX = playerX - SW / 2 + CHAR_W / 2;
-        float maxCam = (float)(fgWorldWidth - SW);
+        double cameraX = playerX - SW / 2 + CHAR_W / 2;
+        double maxCam = (double)(fgWorldWidth - SW);
         if (cameraX < 0) cameraX = 0;
         if (cameraX > maxCam) cameraX = maxCam;
